torch/requirements.txt
torch==1.2.0
torchvision==0.4.0
scikit-image==0.14.5
Pillow==6.1.0
plyfile==0.7
SSIM_PIL==1.0.14
chamferdist==1.0.0
